# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Quran API Go
  description: |-
    Internal RESTful API that serves Al-Quran data (Arabic text, Indonesian & English translations)
    for the Ilmunara super app.
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /surah:
    get:
      tags:
        - Surah
      summary: List all surahs
      description: Get a list of all 114 surahs, optionally filtered by revelation type
      operationId: listSurahs
      parameters:
        - name: type
          in: query
          description: Filter by revelation type
          schema:
            type: string
            enum: [meccan, medinan]
      responses:
        '200':
          description: List of surahs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Surah'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /surah/{id}:
    get:
      tags:
        - Surah
      summary: Get surah by ID
      description: Get detailed information about a specific surah
      operationId: getSurahById
      parameters:
        - name: id
          in: path
          required: true
          description: Surah ID (1-114)
          schema:
            type: integer
            minimum: 1
            maximum: 114
      responses:
        '200':
          description: Surah details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Surah'
        '400':
          description: Invalid surah ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Surah not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /juz:
    get:
      tags:
        - Juz
      summary: List all juz
      description: Get a list of all 30 juz (parts) of the Quran
      operationId: listJuzs
      responses:
        '200':
          description: List of juz
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Juz'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /juz/{number}:
    get:
      tags:
        - Juz
      summary: Get juz by number
      description: Get detailed information about a specific juz
      operationId: getJuzByNumber
      parameters:
        - name: number
          in: path
          required: true
          description: Juz number (1-30)
          schema:
            type: integer
            minimum: 1
            maximum: 30
      responses:
        '200':
          description: Juz details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Juz'
        '400':
          description: Invalid juz number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Juz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /juz/{number}/ayah:
    get:
      tags:
        - Juz
      summary: Get ayahs by juz
      description: Get all ayahs from a specific juz with pagination
      operationId: getAyahsByJuz
      parameters:
        - name: number
          in: path
          required: true
          description: Juz number (1-30)
          schema:
            type: integer
            minimum: 1
            maximum: 30
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: lang
          in: query
          description: Translation language
          schema:
            type: string
            enum: [id, en]
            default: id
      responses:
        '200':
          description: Juz ayahs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JuzAyahsResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Juz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /juz/{number}/surah:
    get:
      tags:
        - Juz
      summary: Get surahs by juz
      description: Get all surahs that appear in a specific juz
      operationId: getSurahsByJuz
      parameters:
        - name: number
          in: path
          required: true
          description: Juz number (1-30)
          schema:
            type: integer
            minimum: 1
            maximum: 30
      responses:
        '200':
          description: List of surahs in the juz
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Surah'
        '400':
          description: Invalid juz number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Juz not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    SuccessResponse:
      type: object
      required:
        - data
        - timestamp
      properties:
        data:
          type: object
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp in UTC
    ErrorResponse:
      type: object
      required:
        - error
        - code
        - timestamp
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - bad request
            - not found
            - internal server error
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp in UTC
...
    Surah:
      type: object
      required:
        - id
        - number
        - name_arabic
        - name_latin
        - name_transliteration
        - number_of_ayahs
        - revelation_type
      properties:
        id:
          type: integer
          minimum: 1
          maximum: 114
          example: 1
        number:
          type: integer
          minimum: 1
          maximum: 114
          example: 1
        name_arabic:
          type: string
          example: الفاتحة
        name_latin:
          type: string
          example: Al-Fatihah
        name_transliteration:
          type: string
          example: Al-Faatihah
        number_of_ayahs:
          type: integer
          minimum: 3
          example: 7
        revelation_type:
          type: string
          enum: [Meccan, Medinan]
          example: Meccan
...
    Juz:
      type: object
      required:
        - id
        - juz_number
        - first_ayah_id
        - last_ayah_id
        - total_ayahs
      properties:
        id:
          type: integer
          example: 1
        juz_number:
          type: integer
          minimum: 1
          maximum: 30
          example: 1
        first_ayah_id:
          type: integer
          example: 1
        last_ayah_id:
          type: integer
          example: 150
        total_ayahs:
          type: integer
          example: 150
    JuzAyahsResponse:
      type: object
      required:
        - juz
        - ayahs
      properties:
        juz:
          type: object
          required:
            - juz_number
            - total_ayahs
          properties:
            juz_number:
              type: integer
              example: 1
            total_ayahs:
              type: integer
              example: 150
        ayahs:
          type: array
          items:
            type: object
            required:
              - id
              - surah_id
              - surah_name
              - number_in_surah
              - text_uthmani
              - translation
              - juz_number
            properties:
              id:
                type: integer
                example: 1
              surah_id:
                type: integer
                example: 1
              surah_name:
                type: string
                example: Al-Fatihah
              number_in_surah:
                type: integer
                example: 1
              text_uthmani:
                type: string
                example: بِسْمِ ٱللَّهِ ٱلرَّحْمَـٰنِ ٱلرَّحِيمِ
              translation:
                type: string
                example: Dengan nama Allah Yang Maha Pengasih, Maha Penyayang.
              juz_number:
                type: integer
                example: 1
...